requirements.txt
python-telegram-bot==20.3
camelot-py==0.10.1
pandas==1.5.3
PyYAML==6.0
pdfminer.six==20221105
openpyxl==3.1.2
numpy==1.23.5
opencv-python-headless==4.5.5.64
PyPDF2==2.12.1